requirements.txt
dbt-bigquery==1.9.1
google-cloud-core==2.4.1
google-cloud-bigquery==3.14.1
google-cloud-storage==2.14.0
pandas==2.2.2
pyarrow==13.0.0
Faker==19.0.1
numpy==1.24.1
